test(projects): consolidate project test suites, docs/CLI/sync parity
Merges projectDetailService.test.ts into projectLifecycle.test.ts; internal
docs verified against final code; ade projects inspect CLI command +
projects.inspectPath RPC; sync host surfaces coded error codes cleanly to
mobile/TUI clients instead of leaking the message prefix.
